Minimize the number of functions declared inside giant "let rec ... and ..." block
I do not understand why yet, but I have found that, for example, my printf wrapper functions did not work if declared with "let rec ... and ...", because the compiler was inferring a specific type for them rather than leaving them polymorphic.

So it seems like it is a best practice to minimize the use of "let rec" to when it is actually required.