Backtrack on trying to have only a minimal number of constructors in the IR.
I realized that when I add floating point, I will need more than CompEQ and CompLT, because, eg, it is not necessarily the case that !(x < y) == (x >= y) if x and y are floats and NaNs are involved.

Also, it's not like I'm going to try to do CSE, so it's not super critical to have everything fully normalized.

So I add the full complement of comparison functions, and LogicalNot and Sub for good measure.