Fix intermittent bug when calling varags functions with floating point arguments, such as printf.
The ABI requires an upper bound of the number of vector registers used to be placed in "al", and we were neglecting to do that.

Instead, when calling printf, "al" happened to contain the low bits of the pointer to the format string, so the behavior of printf was different depending on whether the format string was aligned to a multiple of 256 bytes.