2nd try at parsing type declarations.
Before, I had the wrong precedence of [] vs * in type declarations and I wasn't reversing the nesting order of the usage when calculating the type, so in simple cases, two wrongs made a right.

Now I've swapped the precedence, reversed the nesting order, and added more cases to regression.c.