Move iterative dataflow analysis out of Liveness module into Regalloc, and delete Liveness module.
I realized that building the interference graph does not depend solely on the "live out" set of each instruction, but also whether the instruction is a move or not. Rather than trying to feed that information into the Liveness module, I have simply moved the dataflow analysis into the Regalloc module, which makes for a simpler interface to Regalloc.